+++ suporteEngenhariaUI/Models/CloseStatusApi.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Text.Json.Serialization;
+using System.Threading.Tasks;
+
+namespace suporteEngenhariaUI
+{
+    public class CloseStatusApi
+    {
+        [JsonPropertyName("status")] public required string Status { get; set; }
+        [JsonPropertyName("error")] public string? Error { get; set; }
+    }
+}
